requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
PuLP==2.6.0
python-constraint==1.4.0
Werkzeug==2.0.3
